refactor: ideaMarketEdit.ts에서 type 선언 로직 분리(specializationType.ts)
