+++ src/types/specializationType.ts
@@ -1,0 +1,14 @@
+export type SpecializationType =
+  | 'ADVERTISING_PROMOTION'
+  | 'DESIGN'
+  | 'LESSON'
+  | 'MARKETING'
+  | 'DOCUMENT_WRITING'
+  | 'MEDIA_CONTENT'
+  | 'TRANSLATION_INTERPRETATION'
+  | 'TAX_LAW_LABOR'
+  | 'CUSTOM_PRODUCTION'
+  | 'STARTUP_BUSINESS'
+  | 'FOOD_BEVERAGE'
+  | 'IT_TECH'
+  | 'OTHERS';
